refactor common to utils
